migrate 'range_join_across_single_row' test over to re_query
