+++ crates/re_query/Cargo.toml
@@ -50,6 +50,7 @@ criterion.workspace = true
 itertools = { workspace = true }
 mimalloc.workspace = true
 rand = { workspace = true, features = ["std", "std_rng"] }
+similar-asserts.workspace = true
 
 
 [lib]
